2018-4-13

+++ appinterface/src/main/java/com/emi/roots/entity/AzCheckin.java
@@ -5,7 +5,7 @@ public class AzCheckin {
 
     private Integer uid;
 
-    private Integer createtime;
+    private Long createtime;
 
     public Integer getId() {
         return id;
@@ -23,11 +23,11 @@ public class AzCheckin {
         this.uid = uid;
     }
 
-    public Integer getCreatetime() {
+    public Long getCreatetime() {
         return createtime;
     }
 
-    public void setCreatetime(Integer createtime) {
+    public void setCreatetime(Long createtime) {
         this.createtime = createtime;
     }
 }
