+++ appinterface/src/main/java/com/emi/roots/entity/AzScoreLog.java
@@ -15,7 +15,7 @@ public class AzScoreLog {
 
     private Integer finally_value;
 
-    private Integer create_time;
+    private Long create_time;
 
     private String remark;
 
@@ -79,11 +79,11 @@ public class AzScoreLog {
         this.finally_value = finally_value;
     }
 
-    public Integer getCreate_time() {
+    public Long getCreate_time() {
         return create_time;
     }
 
-    public void setCreate_time(Integer create_time) {
+    public void setCreate_time(Long create_time) {
         this.create_time = create_time;
     }
