+++ appinterface/src/main/java/com/emi/roots/entity/AzSof.java
@@ -5,7 +5,7 @@ public class AzSof {
 
     private Integer uid;
 
-    private Integer create_time;
+    private Long create_time;
 
     private Integer update_time;
 
@@ -43,11 +43,11 @@ public class AzSof {
         this.uid = uid;
     }
 
-    public Integer getCreate_time() {
+    public Long getCreate_time() {
         return create_time;
     }
 
-    public void setCreate_time(Integer create_time) {
+    public void setCreate_time(Long create_time) {
         this.create_time = create_time;
     }
